added support for fractional sorting for product images

+++ src/products/images/dto.rs
@@ -1,5 +1,6 @@
 use crate::products::images::model::ProductImageModel;
 use crate::traits::HasId;
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -9,7 +10,7 @@ pub struct PublicProductImage {
     pub url: String,
     pub alt: String,
     pub is_main: bool,
-    pub sort: i32,
+    pub sort: BigDecimal,
 }
 
 impl HasId for PublicProductImage {
